requirements.txt
streamlit
torch
diffusers
transformers
Pillow
matplotlib
numpy
insightface
onnxruntime
spaces